Compilation fixes
