+++ src/planner/binder/statement/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library_unity(
   duckdb_bind_statement
   OBJECT
+  bind_alter.cpp
   bind_attach.cpp
   bind_call.cpp
   bind_connect.cpp
@@ -24,8 +25,8 @@ add_library_unity(
   bind_relation.cpp
   bind_select.cpp
   bind_set.cpp
-  bind_simple.cpp
   bind_summarize.cpp
+  bind_transaction.cpp
   bind_update.cpp
   bind_update_extensions.cpp
   bind_vacuum.cpp)
